marketplace/requirements.txt
Django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==11.2.1
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.1
dj-database-url==2.1.0 